Fixed old program 2 mistakes

+++ asg3/lyutils.h
@@ -18,7 +18,7 @@ extern int yyin_linenr;
 extern char* yytext;
 extern int yy_flex_debug;
 extern int yydebug;
-extern unsigned long yyleng;
+extern int yyleng;
 
 extern ofstream tok_file;
